Simplified wallet setup workflow to easily migrate to Electrum from web wallet
The "Standard" wallet creation workflow is appropriately verbose for
typical workflows, but to make migration to Electrum simpler from the
web wallet, there is a workflow that makes a lot of basic assumptions
using language that was used before:

1. Create a wallet file name,
2. Choose "Restore from Web Wallet"
3. See the seed dialog, but with the "transaction password" (or
"extended words") and a way to verify that the wallet is properly
derived by adding the first address.
4. It will allow to create the wallet if it is possible to derive a
P2PKH keystore with the first address being the first address derived
from the keystore.

Derivation is intensive for UX purposes, so the derivation to do a
test is delayed until all inputs are ready. If there is a derivation
error, it will inform the user that there was a failure.